Don't load in config before first slack message
